Laravel middleware and profile

+++ backend/laravel/routes/api.php
@@ -20,9 +20,11 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('auth:sanctum')->get('/user/login', function (Request $request) {
-    return $request->user();
-});
+// Route::middleware('auth:')->get('/user/login', function (Request $request) {
+//     return $request->user();
+// });
+
+Route::get('/profile',[UserController::class,'profile'])->name('login'); 
 
 Route::prefix('api/category')->group(function () {
     Route::get('/', [CategoryController::class, 'index']);
@@ -30,7 +32,7 @@ Route::prefix('api/category')->group(function () {
     Route::post('/', [CategoryController::class, 'store']);
     Route::put('/{id}', [CategoryController::class, 'update']);
     Route::delete('/{id}', [CategoryController::class, 'destroy']);
-});
+})->middleware('auth:api');;
 
 Route::prefix('api/table')->group(function () {
     Route::get('/', [TableController::class, 'index']);
@@ -38,7 +40,7 @@ Route::prefix('api/table')->group(function () {
     Route::post('/', [TableController::class, 'store']);
     Route::put('/{id}', [TableController::class, 'update']);
     Route::delete('/{id}', [TableController::class, 'destroy']);
-});
+})->middleware('auth:api');;
 
 Route::prefix('api/dishtype')->group(function () {
     Route::get('/', [DishTypeController::class, 'index']);
@@ -46,7 +48,7 @@ Route::prefix('api/dishtype')->group(function () {
     Route::post('/', [DishTypeController::class, 'store']);
     Route::put('/{id}', [DishTypeController::class, 'update']);
     Route::delete('/{id}', [DishTypeController::class, 'destroy']);
-});
+})->middleware('auth:api');
 
 Route::prefix('api/dish')->group(function () {
     Route::get('/', [DishController::class, 'index']);
@@ -54,14 +56,15 @@ Route::prefix('api/dish')->group(function () {
     Route::post('/', [DishController::class, 'store']);
     Route::put('/{id}', [DishController::class, 'update']);
     Route::delete('/{id}', [DishController::class, 'destroy']);
-});
+})->middleware('auth:api');;
 
 Route::prefix('api/user')->group(function () {
-    Route::get('/', [UserController::class, 'index']);
-    Route::get('/{id}', [UserController::class, 'show']);
-    Route::post('/', [UserController::class, 'store']);
+    Route::get('/', [UserController::class, 'index'])->middleware('auth:api');
+    Route::get('/profile', [UserController::class, 'profile'])->middleware('auth:api');
+    Route::get('/{id}', [UserController::class, 'show'])->middleware('auth:api');
+    Route::post('/', [UserController::class, 'store'])->middleware('auth:api');
     Route::post('/login', [UserController::class, 'login']);
-    Route::put('/{id}', [UserController::class, 'update']);
-    Route::delete('/{id}', [UserController::class, 'destroy']);
+    Route::put('/{id}', [UserController::class, 'update'])->middleware('auth:api');
+    Route::delete('/{id}', [UserController::class, 'destroy'])->middleware('auth:api');
 });
